reorganized user loging and registration methods
ready: login, register, delete user
