+++ backend/matcha/__init__.py
@@ -40,7 +40,7 @@ def create_app(test_config=None):
     from . import db
     db.init_app(app)
 
-    from . import auth
-    app.register_blueprint(auth.bp)
+    from . import user
+    app.register_blueprint(user.bp)
 
     return app
